refactor(clock): keep SystemClock in core as zero-dep default
Devin review on PR #591: HLE-54 moved both SystemClock and FrozenClock to
adapters/clock.py, then 7 Core services did lazy `from atman.adapters.clock
import SystemClock` inside __init__ to avoid a static cross-layer import.
That works but inverts the dependency direction at runtime — Core falling
back to an adapter for its own default.

The fix: SystemClock is a 4-line stdlib wrapper over datetime.now(UTC) with
no external dependency, so it's really a core default, not an adapter.
Move it back to core/clock_impl.py beside ensure_utc; FrozenClock stays in
adapters/clock.py because it is genuinely a test-fixture adapter. Core
services import SystemClock statically from core/clock_impl, dropping the
lazy fallback. e2e scripts that wanted wall-clock behaviour switch their
SystemClock import to core/clock_impl; FrozenClock importers stay on
atman.adapters.clock.

Tests, ruff, ruff format, pyright all green; SYSTEM_MAP (en + ru) updated.
