+++ docs/architecture/SYSTEM_MAP-ru.md
@@ -98,7 +98,7 @@
 |------|------------|
 | `config.py` | Pydantic settings (`EmbeddingSettings`, `LLMSettings`, `MemorySettings`), **`SkillsSettings`** (WP-08 v2: `enabled`, `skills_root`, `auto_pin_threshold_uses`, `auto_pin_threshold_sessions`, `auto_downgrade_sessions`, `min_confidence`), **`OpenAILLMConfig`** (base_url, api_key, model, timeout, max_retries с валидацией ≥1), **`AnthropicLLMConfig`** (api_key, model, max_tokens), фабрика `build_memory_backend()`, **фабрика `build_embedding_adapter()`** (выбор FlagEmbedding/Ollama/Mock backend), `validate_embedding_dimension()` (проверка размерности при старте); по умолчанию: embedding backend=`ollama` с `bge-m3`/1024d (FlagEmbedding backend использует `flag_model="BAAI/bge-m3"` с настройками FP16/batch_size/max_length), LLM=`gemma3:27b-it-qat`, factual memory=`FileBackend`; поддерживает `ATMAN_MEMORY_BACKEND=postgres|file|inmemory`, `EMBEDDING_BACKEND=ollama|flag|mock`; **fallback устаревших переменных окружения**: `OLLAMA_HOST`→`EMBEDDING_OLLAMA_HOST`, `OLLAMA_EMBED_MODEL`→`EMBEDDING_MODEL`, `ATMAN_OLLAMA_BASE_URL`→`LLM_OLLAMA_HOST`, `ATMAN_OLLAMA_MODEL`→`LLM_MODEL`; **opt-2**: `EmbeddingSettings.cache_size` (по умолчанию 4096, 0=отключено) передаётся в оба адаптера |
 | `core/exceptions.py` | `AtmanError`, `GovernanceRejectedError`, `NarrativePersistenceConflictError`, `SessionNotFoundError`, `SessionAlreadyFinishedError`, `TooManyActiveSessionsError` |
-| `core/clock_impl.py` | `ensure_utc(dt)` — чистый нормализатор временной зоны; конкретные `SystemClock` / `FrozenClock` переехали в `adapters/clock.py` (HLE-54) |
+| `core/clock_impl.py` | `SystemClock` (дефолтный `ClockPort`, zero-dep обёртка над stdlib) + `ensure_utc(dt)` — чистый нормализатор временной зоны; `FrozenClock`-фикстура живёт в `adapters/clock.py` (HLE-54) |
 | `core/narrative_write_audit.py` | Хуки аудита коммитов нарратива |
 | `core/reflection_event_audit.py` | Наблюдатели персистенса событий рефлексии |
 | `core/reflection_run_keys.py` | Детерминированные ключи прогонов рефлексии |
@@ -151,7 +151,7 @@
 | `adapters/storage/postgres_pending_human_review.py` (R11.7) | `PendingHumanReviewInbox` | `agent_{N}.pending_human_review`; при enqueue пишет `agent_id` в `context` |
 | `adapters/storage/reflection_store.py` (`ReflectionStore`) | — | PostgreSQL `agent_{N}.reflections` через `AgentSchemaResolver` (без RLS); **переехал сюда из `atman/reflection/store.py` в HLE-53 — старый пакет распущен, persistence-модель живёт в `adapters/storage/reflection_persistence_models.py` и переиспользует канонический `core.models.reflection.ReflectionLevel`** |
 | `adapters/storage/_atomic_write.py` (`write_atomically`) (HLE-49) | — | Общий хелпер tempfile + fsync + chmod + os.replace; используется `FileBackend._save_facts` и `FileStateStore._write_json_atomically`, чтобы контракт атомарной записи жил в одном месте |
-| `adapters/clock.py` (`SystemClock`, `FrozenClock`) (HLE-54) | `ClockPort` | Wall-clock + замороженные часы для тестов; переехали из `core/clock_impl.py`, чтобы Core больше не содержал конкретных реализаций |
+| `adapters/clock.py` (`FrozenClock`) (HLE-54) | `ClockPort` | Замороженные часы для тестов; дефолтный `SystemClock` живёт в `core/clock_impl.py`, так как это zero-dep обёртка над stdlib |
 | `adapters/storage/postgres_agent_schema.py` | — | резолв `agent_id` → схема `agent_{serial_id}` для субъективных Postgres-адаптеров |
 | `adapters/storage/in_memory_reflection_request_queue.py` (R12) | `ReflectionRequestQueue` | идемпотентность в пределах UTC-часа через `agent_driven_run_key(reason, hour)` |
 | `adapters/observability/in_memory_overload_alert_sink.py` (R13) | `ReflectionOverloadAlertSink` | алерты в памяти; падения sink подавляются, чтобы монитор не валил вызывающего |
