+++ docs/architecture/SYSTEM_MAP.md
@@ -106,7 +106,7 @@ All paths are absolute relative to the repository root.
 |------|---------|
 | `config.py` | Pydantic settings (`EmbeddingSettings`, `LLMSettings`, `MemorySettings`), **`SkillsSettings`** (WP-08 v2: `enabled`, `skills_root`, `auto_pin_threshold_uses`, `auto_pin_threshold_sessions`, `auto_downgrade_sessions`, `min_confidence`), **`OpenAILLMConfig`** (base_url, api_key, model, timeout, max_retries with validation ≥1), **`AnthropicLLMConfig`** (api_key, model, max_tokens), `build_memory_backend()` factory, **`build_embedding_adapter()` factory** (selects FlagEmbedding/Ollama/Mock backend), `validate_embedding_dimension()` (startup dimension check); defaults: embedding backend=`ollama` with `bge-m3`/1024d (FlagEmbedding backend uses `flag_model="BAAI/bge-m3"`, with FP16/batch_size/max_length settings), LLM=`gemma3:27b-it-qat`, factual memory=`FileBackend`; supports `ATMAN_MEMORY_BACKEND=postgres|file|inmemory`, `EMBEDDING_BACKEND=ollama|flag|mock`; **legacy env var fallback**: `OLLAMA_HOST`→`EMBEDDING_OLLAMA_HOST`, `OLLAMA_EMBED_MODEL`→`EMBEDDING_MODEL`, `ATMAN_OLLAMA_BASE_URL`→`LLM_OLLAMA_HOST`, `ATMAN_OLLAMA_MODEL`→`LLM_MODEL`; **opt-2**: `EmbeddingSettings.cache_size` (default 4096, 0=disabled) passed to both FlagEmbedding and Ollama adapters |
 | `core/exceptions.py` | `AtmanError`, `GovernanceRejectedError`, `NarrativePersistenceConflictError`, `SessionNotFoundError`, `SessionAlreadyFinishedError`, `TooManyActiveSessionsError` |
-| `core/clock_impl.py` | `ensure_utc(dt)` — pure timezone normaliser; concrete `SystemClock` / `FrozenClock` moved to `adapters/clock.py` (HLE-54) |
+| `core/clock_impl.py` | `SystemClock` (default `ClockPort`, zero-dep stdlib wrapper) + `ensure_utc(dt)` pure timezone normaliser; `FrozenClock` test-fixture lives in `adapters/clock.py` (HLE-54) |
 | `core/narrative_write_audit.py` | Narrative commit audit hooks |
 | `core/reflection_event_audit.py` | Reflection event persistence observers |
 | `core/reflection_run_keys.py` | Deterministic reflection run keys |
@@ -147,7 +147,7 @@ All paths are absolute relative to the repository root.
 | `adapters/storage/postgres_pending_human_review.py` (R11.7) | `PendingHumanReviewInbox` | `agent_{N}.pending_human_review`; enqueues `agent_id` into `context` |
 | `adapters/storage/reflection_store.py` (`ReflectionStore`) | — | PostgreSQL `agent_{N}.reflections` via `AgentSchemaResolver` (no RLS); **moved here from `atman/reflection/store.py` in HLE-53 — old package dissolved, persistence model lives in `adapters/storage/reflection_persistence_models.py` and re-uses the canonical `core.models.reflection.ReflectionLevel`** |
 | `adapters/storage/_atomic_write.py` (`write_atomically`) (HLE-49) | — | Shared tempfile + fsync + chmod + os.replace helper; backs `FileBackend._save_facts` and `FileStateStore._write_json_atomically` so the atomic-write contract is single-sourced |
-| `adapters/clock.py` (`SystemClock`, `FrozenClock`) (HLE-54) | `ClockPort` | Wall-clock + test-only frozen clock; moved out of `core/clock_impl.py` so Core no longer ships concrete implementations |
+| `adapters/clock.py` (`FrozenClock`) (HLE-54) | `ClockPort` | Test-fixture frozen clock; the default `SystemClock` lives in `core/clock_impl.py` because it is a zero-dep stdlib wrapper |
 | `adapters/storage/postgres_agent_schema.py` | — | `agent_id` → `agent_{serial_id}` schema resolution for subjective Postgres adapters |
 | `adapters/storage/in_memory_reflection_request_queue.py` (R12) | `ReflectionRequestQueue` | idempotent within UTC hour bucket via `agent_driven_run_key(reason, hour)` |
 | `adapters/observability/in_memory_overload_alert_sink.py` (R13) | `ReflectionOverloadAlertSink` | captures alerts in-memory; sink failures suppressed so monitor cannot crash callers |
