+++ e2e/full_loop.py
@@ -33,7 +33,7 @@ from datetime import UTC, datetime, timedelta
 from pathlib import Path
 from uuid import UUID, uuid4
 
-from atman.adapters.clock import FrozenClock, SystemClock
+from atman.adapters.clock import FrozenClock
 from atman.adapters.reflection.state_store_session_repository import (
     StateStoreSessionRepository,
 )
@@ -43,6 +43,7 @@ from atman.adapters.storage.in_memory_reflection_store import (
     InMemoryPatternStore,
     InMemoryReflectionEventStore,
 )
+from atman.core.clock_impl import SystemClock
 from atman.core.models import (
     CoreValue,
     Goal,
