+++ e2e/scenarios/session_lifecycle_interrupt.py
@@ -36,8 +36,8 @@ import tempfile
 from pathlib import Path
 from uuid import UUID, uuid4
 
-from atman.adapters.clock import SystemClock
 from atman.adapters.storage.file_state_store import FileStateStore
+from atman.core.clock_impl import SystemClock
 from atman.core.models import (
     CoreValue,
     Goal,
